main: split out 'shells' module for init
This moves giant const strings off into functions where they look less
egregious.
