+++ src/lib.rs
@@ -1,3 +1,3 @@
 // This token 'lib' file is here for reusing code between the bin and integ tests.
 
-pub mod supported_shells;
+pub mod shells;
