--- src/supported_shells.rs
@@ -1,1 +1,0 @@
-pub const SUPPORTED_SHELLS: [&str; 2] = ["zsh", "bash"];
